requirements.txt
Django==2.2
django-ckeditor==5.7.1
django-debug-toolbar==2.1
django-js-asset==1.2.2
django-redis==4.10.0
django-tinymce==2.8.0
Pillow==6.0.0
PyMySQL==0.9.3
pytz==2019.1
redis==3.3.11
sqlparse==0.3.0
